refactor: move humanizedDbMap to db utils, remove test
