+++ ui/apps/everest/src/pages/databases/create-db-button/create-db-button.tsx
@@ -19,9 +19,9 @@ import { ArrowDropDownIcon } from '@mui/x-date-pickers/icons';
 import { Messages } from '../dbClusterView.messages';
 import { useDBEnginesForDbEngineTypes } from 'hooks';
 import { dbEngineToDbType } from '@percona/utils';
-import { humanizeDbType } from 'pages/database-form/database-form.utils';
 import { Link, useNavigate } from 'react-router-dom';
 import { useNamespacePermissionsForResource } from 'hooks/rbac';
+import { humanizeDbType } from 'utils/db';
 
 export const CreateDbButton = () => {
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
