+++ ui/apps/everest/src/utils/db.tsx
@@ -298,3 +298,11 @@ export const setDbClusterRestart = (dbCluster: DbCluster) => ({
     },
   },
 });
+
+const humanizedDbMap: Record<DbType, string> = {
+  [DbType.Postresql]: 'PostgreSQL',
+  [DbType.Mongo]: 'MongoDB',
+  [DbType.Mysql]: 'MySQL',
+};
+
+export const humanizeDbType = (type: DbType): string => humanizedDbMap[type];
